Add script to only compile models
